🧹 Refactor daily-run.js by extracting generator functions
