+++ marketing/lib/content.js
@@ -1,0 +1,146 @@
+function generateBlogPost(keyword, template) {
+  const titles = {
+    "how-to": `How to Master ${keyword}: The Complete Guide`,
+    guide: `The Definitive Guide to ${keyword} in 2026`,
+    vs: `${keyword} vs Manual: Which is Right for You?`,
+  };
+
+  const title = titles[template] || titles.guide;
+
+  return `---
+title: "${title}"
+slug: "${keyword.replace(/\s+/g, "-")}"
+excerpt: "Learn how to ${keyword.toLowerCase()} with this comprehensive guide. Save time and reduce errors with proven strategies."
+keywords: ["${keyword}", "reconciliation", "automation"]
+readingTime: 8
+metaDescription: "Master ${keyword} with our comprehensive guide. Learn best practices, tools, and strategies for 2026."
+date: "${new Date().toISOString()}"
+---
+
+# ${title}
+
+If you're struggling with ${keyword}, you're not alone. Here's a complete guide to solving it.
+
+## The Problem with Manual ${keyword}
+
+Most teams approach ${keyword} reactively. They wait for discrepancies to surface. This creates a constant firefighting mode.
+
+## Understanding ${keyword} Fundamentals
+
+At its core, ${keyword} is about ensuring your financial records match across systems. Sounds simple, but at scale, complexity explodes.
+
+## Step 1: Set Up Your Data Sources
+
+Connect all your payment processors, bank accounts, and internal systems. API-first beats CSV exports every time.
+
+## Step 2: Define Your Matching Rules
+
+What constitutes a "match"? Same amount? Same date? Same customer? Define this once, apply it consistently.
+
+## Step 3: Automate the Reconciliation
+
+With the right tooling, 95% of transactions match automatically. You only review exceptions.
+
+## Common ${keyword} Pitfalls to Avoid
+
+1. **Waiting until month-end** - Reconcile daily, not monthly
+2. **Over-engineering matching rules** - Start simple, iterate
+3. **Ignoring edge cases** - Handle refunds, chargebacks, FX separately
+
+${keyword} is essential for accurate financial reporting. With the right approach, you can save hours every month.
+
+Ready to automate ${keyword}? [Try Settler free for 14 days](https://settler.dev/demo).
+`;
+}
+
+function generateSocialContent(idea) {
+  return `CONTENT MULTIPLIER RESULTS
+=========================
+
+Original Idea: ${idea}
+Generated: 8 pieces
+
+---
+
+TWITTER
+-------
+
+[THREAD 1]
+Best time: 9:00 AM EST
+
+I analyzed 100 companies' reconciliation workflows.
+
+Here's what the top 10% do differently:
+
+🧵
+
+1/ Most teams approach reconciliation reactively. They wait for discrepancies to surface. Proactive monitoring changes everything.
+
+2/ The #1 mistake? Trying to reconcile at the transaction level. Aggregate first, then drill down. 10x faster.
+
+3/ APIs changed the game. Real-time data means real-time reconciliation. No more waiting for CSV exports.
+
+---
+
+[TWEET 2]
+Best time: 12:00 PM EST
+
+💡 ${idea}
+
+What if I told you this could be automated in 5 minutes?
+
+→ settler.dev/demo
+
+---
+
+LINKEDIN
+========
+
+[POST 1]
+Best time: 8:00 AM EST
+Hashtags: #FinTech #Automation #FinanceOps #SaaS
+
+${idea}
+
+Sounds simple, right?
+
+But 90% of companies are still doing this manually.
+
+Here's why that needs to change:
+
+[Link to full post]
+
+---
+
+REDDIT
+======
+
+[POST 1]
+Best time: 2:00 PM EST
+
+[Discussion] ${idea} - What's your experience?
+
+I've been working with teams on reconciliation automation, and this is the #1 thing that moves the needle.
+
+Curious: Are you doing this manually or have you automated? What's worked for you?
+
+---
+
+HACKER NEWS
+===========
+
+[COMMENT 1]
+Best time: On relevant thread
+
+This resonates with our experience. We found that ${idea} was the turning point for our finance ops.
+
+The key insight: most teams try to solve reconciliation with more people, when they should be solving it with better tooling.
+
+We've seen companies cut reconciliation time by 90% with the right automation.
+`;
+}
+
+module.exports = {
+  generateBlogPost,
+  generateSocialContent,
+};
